Rework mathematics cog
