+++ src/math_constants.py
@@ -1,0 +1,19 @@
+import math
+from typing import Union
+import numpy as np
+
+numeric = Union[int, float, complex, np.number]
+
+pi: float = math.pi
+alpha: float = 2.502907875095892822283902873218
+delta: float = 4.669201609102990671853203821578
+theta: float = 1.30637788386308069046861449260260571
+tau: float = math.tau
+phi: float = 1.6180339887498948482
+gamma: float = 0.57721566490153286060651209008240243104215933593992
+lambda_var: float = 1.303577269034
+psi: float = 3.35988566624317755317201130291892717
+rho: float = 1.32471795724474602596090885447809734
+e: float = math.e
+i: complex = complex(0,1)
+inf: float = math.inf
